chore(config): make obsolete check from QueryConfig to Config
