+++ src/query/storages/system/src/configs_table.rs
@@ -17,8 +17,8 @@ use std::sync::Arc;
 use common_base::base::mask_string;
 use common_catalog::table::Table;
 use common_catalog::table_context::TableContext;
+use common_config::Config;
 use common_config::GlobalConfig;
-use common_config::QueryConfig;
 use common_exception::Result;
 use common_expression::types::StringType;
 use common_expression::utils::FromData;
@@ -285,7 +285,7 @@ impl ConfigsTable {
     fn remove_obsolete_query_configs(config_json: JsonValue) -> JsonValue {
         match config_json {
             Value::Object(mut config_json_obj) => {
-                for key in QueryConfig::obsoleted_option_keys().iter() {
+                for key in Config::obsoleted_option_keys().iter() {
                     config_json_obj.remove(*key);
                 }
                 JsonValue::Object(config_json_obj)
